Splits uploadvelden.class.php
