+++ lib/view/formulier/uploadvelden/RequiredFileField.class.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * RequiredFileField.class.php
+ *
+ * @author C.S.R. Delft <pubcie@csrdelft.nl>
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see FileField
+ */
+class RequiredFileField extends FileField {
+
+	public $required = true;
+
+}
