+++ lib/view/formulier/uploadvelden/RequiredImageField.class.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * RequiredImageField.class.php
+ *
+ * @author C.S.R. Delft <pubcie@csrdelft.nl>
+ * @author P.W.G. Brussee <brussee@live.nl>
+ * @author G.J.W. Oolbekkink <g.j.w.oolbekkink@gmail.com>
+ * @date 30/03/2017
+ *
+ * @see ImageField
+ */
+class RequiredImageField extends ImageField {
+
+	public $required = true;
+
+}
